+++ CHANGELOG.md
@@ -47,6 +47,7 @@
   - Deprecate old-style hardware config dicts and restrict input of `hardware_compile` to a full `CompilationConfig`. (!1002)
 - Operations 
   - Introduce `LongTimeTrace` acquisition protocol. (!958)
+  - Small adjustments to `LongTimeTrace` and tests. (!1014)
 
 ## 0.20.1 (2024-05-01)